+++ src/styles/index.ts
@@ -1,0 +1,2 @@
+export { styled } from "./stitchesConfig";
+export { globalStyles } from "./globalStyles";